Start work on adding a comments thread
Unforutnately I'm having trouble getting the child resources component
to detect box state changes in its parent. I've done it before such that
the props of a child update if the parent passed the prop as a state
var. It's not working yet though so I just added some fake comment data.

For the future, we can't use Aragon's expandable table view component as
3box components don't display properly in the rows. 3Box has its own
comment component so we'll use that as a hideable div most likely.