Submissions sent to new proposal telegram endpoint
A credential is created unsigned and sent to the backend to be forwarded
to the proposals telegram channel. The credential doesn't include the
submitter's information, only the resource's info in case the submitter
is submitting on behalf of someone else. The submitter's contact
information is included in the Telegram message so that we can contact
them.

TODO: update the server URL when it's available at researchcollective.io